CMemoryStream 기능 추가
- ReadFromStream() 메소드 추가
  - Buffer 에 읽기 I/O 없이 참조만 하고, 포지션만 이동
  - stream 내 문자열 버퍼등을 읽어들일때 불필요한 복사를 제거

+++ src/RegistryUtil.cpp
@@ -345,7 +345,7 @@ RUSetMultiString(
 							   NULL,
 							   REG_MULTI_SZ,
 							   (LPBYTE)strm.GetMemory(),
-							   strm.GetSize());
+							   (DWORD)strm.GetSize());
 	if (ERROR_SUCCESS != ret)
 	{
 		//log_err "RegSetValueExW(%ws) failed, ret = %u",
